// cgocore-ui/src/app/features/models/reference/localization/views/full/StationFullView.ts
import { StationForm } from 'src/app/features/models/reference/localization/forms/StationForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class StationFullView {

id: number;
form: StationForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}